# chapter-09/spot-toolkit/docs/openapi.yaml
openapi: 3.1.0
info:
  title: SPOT Toolkit API
  description: |
    RESTful API for the SPOT (Structured Prompt Output Toolkit) content generation system.

    SPOT is an AI-powered content generation tool focused on reliability, monitoring, and evaluation capabilities.
  version: '0.4.0'
  contact:
    name: Chris Minnick
    url: https://github.com/chrisminnick/spot-toolkit
  license:
    name: MIT
    url: https://github.com/chrisminnick/spot-toolkit/blob/main/LICENSE
servers:
  - url: http://localhost:8000
    description: Local development server
...
paths:
...
  /api/v1/templates/validate:
    post:
      tags: [templates]
      summary: Validate templates
      description: Validate all prompt templates
      responses:
        '200':
          description: All templates valid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationResponse'
        '400':
          description: Some templates invalid
...
  /api/v1/style/check:
    post:
      tags: [style]
      summary: Check style compliance
      description: Validate content against style pack rules
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StyleCheckRequest'
      responses:
        '200':
          description: Style check completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StyleCheckResponse'
...
components:
  schemas:
...
    ValidationResponse:
      type: object
      properties:
        validation:
          type: array
          items:
            type: object
            properties:
              templateId:
                type: string
              status:
                type: string
                enum: [valid, invalid]
        summary:
          type: object
          properties:
            total:
              type: integer
            valid:
              type: integer
            invalid:
              type: integer
...
    StyleCheckRequest:
      type: object
      required: [content]
      properties:
        content:
          type: string
    StyleCheckResponse:
      type: object
      properties:
        violations:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
              term:
                type: string
              message:
                type: string
        compliant:
          type: boolean
        stylepack:
          type: object
...